Move log_time to iris
